protobuf: Update Referee message to latest version
Some C++ changes are necessary since the old game event format is now
marked reserved instead of just deprecated, so we can't copy it over
anymore. This hinders our ability to reconstruct vision logs from old
(<= 2018) log files, but this wasn't even completely possible before,
since we only log the time when we received vision packets since 2021
and we don't log the time when we receive Referee packets at all.

Also propagate some integer promotions. Regarding the world state and
logfiles, these should be compatible since on the wire sint32/sint64 and
int32/int64 are the same.

The reflection in the refereestatuswidget must be updated as not to
cause any crashes due to to access via GetMessage

+++ src/protobuf/include/protobuf/ssl_referee.h
@@ -40,7 +40,7 @@ private:
     qint64 m_gameStateChangeTime;
     quint32 m_refereeCounter = 0;
     SSL_Referee::Command m_lastCommand = SSL_Referee::HALT;
-    qint32 m_remainingActionTime = 0;
+    qint64 m_remainingActionTime = 0;
 };
 
 #endif // SSL_REFEREE_H
